+++ CHANGES.rst
@@ -4,7 +4,10 @@ Changelog
 6.2.15 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Add design.plone.contenttypes.behavior.news_base behavior (news metadata, without blocks)
+  [mamico]
+- Add folder creation (multimedia + allegati) for "ComunicatiStampa" CT (if exists)
+  [mamico]
 
 
 6.2.14 (2024-07-11)