+++ CHANGES.rst
@@ -8,7 +8,9 @@ Changelog
   [mamico]
 - Add folder creation (multimedia + allegati) for "ComunicatiStampa" CT (if exists)
   [mamico]
-
+- Add "Emolumenti a carico della finanza pubblica" to Persona
+  Add "Dichiarazioni di insussistenza e incompatibilità" to Persona
+  [lucabel]  
 
 6.2.14 (2024-07-11)
 -------------------