+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 6.2.15 (unreleased)
 -------------------
 
+- Add design.plone.contenttypes.behavior.news_base behavior (news metadata, without blocks)
+  [mamico]
+- Add folder creation (multimedia + allegati) for "ComunicatiStampa" CT (if exists)
+  [mamico]
 - Add "Emolumenti a carico della finanza pubblica" to Persona
   Add "Dichiarazioni di insussistenza e incompatibilità" to Persona
   [lucabel]